sockets.c: Use size_t for iovec length
This change updates the lwip socket read/write ops to use unsigned
size_t for the length of the passed-in iovec.

The internal lwip read/write functions still use int, and simply package
it into struct msghdr, which is size_t once again.
If overflow is a concern we should fix it with a build patch.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>